trying to figure out how to server a js module...
