+++ public/index.html
@@ -242,15 +242,15 @@
 <script src="../socket.io/socket.io.js"></script>
 <script src="scripts/jquery/jquery-3.7.1.min.js"></script>
 <script src="scripts/jquery/jquery-ui-1.13.3.min.js"></script>
-<script src="scripts/escMenu.js"></script>
-<script type="module" src="scripts/attack.js"></script>
-<script type="module" src="scripts/AI.js"></script>
-<script src="scripts/fps.js"></script>
-<script src="scripts/pack_handler.js"></script>
+
 <script type="module" src="scripts/elementsController.js"></script>
+
+<script src="scripts/escMenu.js"></script>
+<!-- <script src="scripts/fps.js"></script> -->
+<!-- <script src="scripts/pack_handler.js"></script> -->
 <script src="scripts/load.js"></script>
-<script src="scripts/fireworks.js"></script>
-<script src="scripts/window_focus.js"></script>
+<!-- <script src="scripts/fireworks.js"></script> -->
+<!-- <script src="scripts/window_focus.js"></script> -->
 <script>
     function showGameMenu() {
         $('#gamemenu').show();
